activities tab is a shared component in pages
